Store procedures

+++ BD/dbo/Tables/Clientes.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[Clientes] (
     [ClientesID]       INT           IDENTITY (1, 1) NOT NULL,
     [ClientesCode]     VARCHAR (5)   NULL,
-    [Cedula]           INT           NULL,
+    [Cedula]           BIGINT        NULL,
     [RazonSocial]      VARCHAR (100) NULL,
     [Empresa]          VARCHAR (100) NULL,
     [ClientesRol]      VARCHAR (100) NULL,
@@ -11,6 +11,4 @@
     [ClientesStatus]   BIT           NULL,
     CONSTRAINT [PK__Clientes__E601B8EE0E586EB1] PRIMARY KEY CLUSTERED ([ClientesID] ASC)
 );
-
-
